Add unit test for NewNodeServer() and move elevated test into a different file
